# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Docker API",
    "description": "A production-ready Go REST API that exposes a simplified wrapper around the Docker Engine API. Supports container lifecycle management and resource metrics.",
    "version": "1.1.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/example/docker-api"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/images": {
      "get": {
        "summary": "List images",
        "description": "Returns all Docker images on the host.",
        "operationId": "listImages",
        "tags": ["Images"],
        "responses": {
          "200": {
            "description": "List of images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ImageSummary" }
                }
              }
            }
          },
          "500": {
            "description": "Failed to list images",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/images/pull": {
      "post": {
        "summary": "Pull image",
        "description": "Pulls an image from a registry. Query param: image=name:tag.",
        "operationId": "pullImage",
        "tags": ["Images"],
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "required": true,
            "description": "Image reference (e.g. alpine:latest)",
            "schema": { "type": "string", "example": "alpine:latest" }
          }
        ],
        "responses": {
          "200": {
            "description": "Image pulled successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PullResponse" }
              }
            }
          },
          "400": {
            "description": "Missing image query parameter",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to pull image",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/images/{id}": {
      "get": {
        "summary": "Inspect image",
        "description": "Returns inspection details for an image.",
        "operationId": "inspectImage",
        "tags": ["Images"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Image ID (full or short) or name:tag",
            "schema": { "type": "string", "example": "abc123def456" }
          }
        ],
        "responses": {
          "200": {
            "description": "Image inspection (types.ImageInspect)",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "400": {
            "description": "Missing or invalid image ID",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to inspect image",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "delete": {
        "summary": "Remove image",
        "description": "Removes an image. Use force=1 or force=true to force removal.",
        "operationId": "removeImage",
        "tags": ["Images"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Image ID (full or short) or name:tag",
            "schema": { "type": "string", "example": "abc123def456" }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Force removal",
            "schema": { "type": "string", "example": "1" }
          }
        ],
        "responses": {
          "200": {
            "description": "Image removed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OperationResponse" },
                "example": { "id": "abc123def456", "status": "removed" }
              }
            }
          },
          "400": {
            "description": "Missing or invalid image ID",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to remove image",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/containers": {
      "get": {
        "summary": "List containers",
        "description": "Returns all Docker containers, including stopped ones.",
        "operationId": "listContainers",
        "tags": ["Containers"],
        "responses": {
          "200": {
            "description": "List of containers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Container" }
                }
              }
            }
          },
          "500": {
            "description": "Failed to communicate with the Docker daemon",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/containers/start/{id}": {
      "post": {
        "summary": "Start a container",
        "description": "Starts a stopped Docker container identified by its ID or name.",
        "operationId": "startContainer",
        "tags": ["Containers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Container ID (full or short) or name",
            "schema": { "type": "string", "example": "abc123def456" }
          }
        ],
        "responses": {
          "200": {
            "description": "Container started successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OperationResponse" },
                "example": { "id": "abc123def456", "status": "started" }
              }
            }
          },
          "400": {
            "description": "Missing or invalid container ID",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Failed to start the container",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/containers/stop/{id}": {
      "post": {
        "summary": "Stop a container",
        "description": "Gracefully stops a running Docker container identified by its ID or name.",
        "operationId": "stopContainer",
        "tags": ["Containers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Container ID (full or short) or name",
            "schema": { "type": "string", "example": "abc123def456" }
          }
        ],
        "responses": {
          "200": {
            "description": "Container stopped successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OperationResponse" },
                "example": { "id": "abc123def456", "status": "stopped" }
              }
            }
          },
          "400": {
            "description": "Missing or invalid container ID",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Failed to stop the container",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/containers/restart/{id}": {
      "post": {
        "summary": "Restart a container",
        "description": "Restarts a running Docker container identified by its ID or name.",
        "operationId": "restartContainer",
        "tags": ["Containers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Container ID (full or short) or name",
            "schema": { "type": "string", "example": "abc123def456" }
          }
        ],
        "responses": {
          "200": {
            "description": "Container restarted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OperationResponse" },
                "example": { "id": "abc123def456", "status": "restarted" }
              }
            }
          },
          "400": {
            "description": "Missing or invalid container ID",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to restart the container",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/containers/{id}": {
      "get": {
        "summary": "Inspect a container",
        "description": "Returns full inspection details for a container.",
        "operationId": "inspectContainer",
        "tags": ["Containers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Container ID (full or short) or name",
            "schema": { "type": "string", "example": "abc123def456" }
          }
        ],
        "responses": {
          "200": {
            "description": "Container inspection (Docker ContainerJSON)",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "400": {
            "description": "Missing or invalid container ID",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to inspect container",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "delete": {
        "summary": "Remove a container",
        "description": "Removes a container. Use force=1 or force=true to remove a running container.",
        "operationId": "removeContainer",
        "tags": ["Containers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Container ID (full or short) or name",
            "schema": { "type": "string", "example": "abc123def456" }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Force removal (e.g. running container)",
            "schema": { "type": "string", "example": "1" }
          }
        ],
        "responses": {
          "200": {
            "description": "Container removed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OperationResponse" },
                "example": { "id": "abc123def456", "status": "removed" }
              }
            }
          },
          "400": {
            "description": "Missing or invalid container ID",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to remove container",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/containers/{id}/logs": {
      "get": {
        "summary": "Get container logs",
        "description": "Returns stdout and stderr of a container (non-streaming).",
        "operationId": "getContainerLogs",
        "tags": ["Containers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Container ID (full or short) or name",
            "schema": { "type": "string", "example": "abc123def456" }
          },
          {
            "name": "tail",
            "in": "query",
            "required": false,
            "description": "Number of lines to show from the end",
            "schema": { "type": "string", "example": "100" }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "Show logs since (timestamp or duration)",
            "schema": { "type": "string" }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "description": "Show logs until (timestamp or duration)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Plain text logs (stdout and stderr combined)",
            "content": {
              "text/plain": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": {
            "description": "Missing or invalid container ID",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Failed to get container logs",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Container": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Full 64-character container ID",
            "example": "abc123def456789abc123def456789abc123def456789abc123def456789abcd"
          },
          "names": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Container names (Docker prefixes each with /)",
            "example": ["/my-container"]
          },
          "image": {
            "type": "string",
            "description": "Image name and tag",
            "example": "nginx:latest"
          },
          "status": {
            "type": "string",
            "description": "Human-readable status string from Docker",
            "example": "Up 2 hours"
          },
          "state": {
            "type": "string",
            "description": "Low-level container state",
            "enum": ["created", "restarting", "running", "removing", "paused", "exited", "dead"],
            "example": "running"
          },
          "labels": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Key/value labels attached to the container"
          },
          "created": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of container creation",
            "example": 1699000000
          }
        }
      },
      "OperationResponse": {
        "type": "object",
        "required": ["id", "status"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Container ID or name echoed from the request",
            "example": "abc123def456"
          },
          "status": {
            "type": "string",
            "description": "Result of the operation",
            "enum": ["started", "stopped", "restarted", "removed"],
            "example": "started"
          }
        }
      },
...
      "ImageSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Image ID" },
          "repo_tags": { "type": "array", "items": { "type": "string" }, "description": "Repository tags" },
          "size": { "type": "integer", "format": "int64", "description": "Size in bytes" },
          "created": { "type": "integer", "format": "int64", "description": "Creation timestamp" },
          "containers": { "type": "integer", "description": "Number of containers using this image" }
        }
      },
      "PullResponse": {
        "type": "object",
        "required": ["status", "image"],
        "properties": {
          "status": { "type": "string", "enum": ["pulled"], "example": "pulled" },
          "image": { "type": "string", "description": "Image reference that was pulled", "example": "alpine:latest" }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "failed to start container"
          }
        }
      }
    }
  }
}